Preparing for 64bits + fixing warnings

+++ OpenKit/OKAuth.h
@@ -2,7 +2,7 @@
 //  OKAuth.h
 //  OpenKit
 //
-//  Created by Manu Mtz-Almeida
+//  Created by Manu Martinez-Almeida
 //  Copyright (c) 2013 OpenKit. All rights reserved.
 //
 
@@ -14,7 +14,7 @@
 
 @interface OKAuthProvider : NSObject
 
-@property(nonatomic, readwrite) int priority;
+@property(nonatomic, readwrite) NSInteger priority;
 @property(nonatomic, readonly) NSString *serviceName;
